+++ ToDoProject/src/main/java/SoftwareEngineering/ToDoProject/TaskObserver.java
@@ -1,5 +1,5 @@
 package SoftwareEngineering.ToDoProject;
 
 public interface TaskObserver {
-    void update(long chosenTaskId);
+    void update(int chosenTaskId);
 }
